Parallel panel LED support
Status LEDs for heaters and fan when using a parallel panel. Default
configuration designed with red/green rotary encoder in mind, but easily
extensible to any combination of LEDs.

Space for this was already allotted in the code, just filled in the
placeholder. Tested with an Arduino Pro Mini and a Sparkfun Red/Green
rotary encoder, no obvious issues were found.